Corección de errores, sintaxis, simplificacion de codigo, ortografia, etc

+++ app/Http/Livewire/Exercises/CreateExercise.php
@@ -34,15 +34,17 @@ class CreateExercise extends Component
         'muscle_group' => 'required',
         // 'type' => 'required',
         'equipment' => 'required',
-        // 'media' => 'required',
+        'media' => 'required',
     ];
 
     public function save()
     {
         $this->validate();
-        $image = Cloudinary::upload($this->media->getRealPath(), ['folder' => 'gym/exercises']);
-        $public_id = $image->getPublicId();
-        $url = $image->getSecurePath();
+        if($this->media){
+            $image = Cloudinary::upload($this->media->getRealPath(), ['folder' => 'gym/exercises']);
+            $public_id = $image->getPublicId();
+            $url = $image->getSecurePath();
+        }
 
         $exercise = Exercise::create([
             'name' => $this->name,
@@ -50,12 +52,13 @@ class CreateExercise extends Component
             'muscle_group' => $this->muscle_group,
             // 'type' => $this->type,
             'equipment' => $this->equipment,
-            'media' => $url,
-            'public_id_media' => $public_id
+            'media' => $url ?? null,
+            'public_id_media' => $public_id ?? null
         ]);
 
         // Eliminamos la imagen temporal
-        File::delete($this->media->getRealPath());
+        if($this->media)
+            File::delete($this->media->getRealPath());
 
         $this->reset(['open', 'name', 'description', 'muscle_group', 'equipment', 'media']);
         $this->emitTo('exercises.show-exercises', 'render');
